requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anthropic
anyio
astunparse
attrs
bert-score
cachetools
certifi
charset-normalizer
colorlog
contourpy
cycler
datasets
dill
distro
filelock
flatbuffers
fonttools
frozenlist
fsspec
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-language
google-genai
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
hf-xet
httpcore
httplib2
httpx
huggingface-hub
idna
Jinja2
jiter
joblib
jsonpatch
jsonpointer
keras
kiwisolver
langchain
langchain-core
langchain-text-splitters
langsmith
libclang
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
mpmath
multidict
multiprocess
namex
networkx
numpy
openai
opt_einsum
optree
optuna
orjson
packaging
pandas
pillow
propcache
proto-plus
protobuf
psycopg2-binary
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
requests-toolbelt
rich
rsa
safetensors
schedule
scikit-learn
scipy
seaborn
sentence-transformers
sentencepiece
setuptools
six
sniffio
SQLAlchemy
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
tf-slim
tf_keras
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
tzdata
uqlm
uritemplate
urllib3
websockets
Werkzeug
wheel
wrapt
xxhash
yarl
zstandard

## pip install -r requirements.txt
## source .venv/bin/activate
